Fix compiler warning
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@25711 f0d54148-0727-0410-94bb-9a71ac55c965

+++ ogr/ogrsf_frmts/vfk/vfkdatablocksqlite.cpp
@@ -488,7 +488,7 @@ int VFKDataBlockSQLite::LoadGeometryPolygon()
             nCount++;
         }
         CPLDebug("OGR-VFK", "%s: fid = %ld nlines = %d -> nrings = %d", m_pszName,
-                 iFID, nLines, poRingList.size());
+                 iFID, (int)nLines, (int)poRingList.size());
 
         if (poLineList.size() > 0) {
             CPLDebug("OGR-VFK",
